+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/model/ElasticSearchExtension.java
@@ -40,6 +40,8 @@ public class ElasticSearchExtension {
 
     private final boolean isKeyword;
 
+    private final boolean isRouting;
+
     public boolean needMatchQuery() {
         return analyzer != null;
     }